move graph id parser to utils
