+++ src/command/graph/push.rs
@@ -7,8 +7,7 @@ use rover_client::query::schema::push;
 use crate::client::get_studio_client;
 use crate::command::RoverStdout;
 use crate::utils::loaders::load_schema_from_flag;
-use crate::utils::parsers::{parse_schema_source, SchemaSource};
-use crate::parsers::parse_graph_id;
+use crate::utils::parsers::{parse_graph_id, parse_schema_source, SchemaSource};
 
 #[derive(Debug, Serialize, StructOpt)]
 pub struct Push {
