+++ src/lib.rs
@@ -1,8 +1,6 @@
 pub mod cli;
 pub(crate) mod client;
 pub mod command;
-mod parsers;
 mod stringify;
 mod telemetry;
 mod utils;
-mod parsers;
